refactoring factories
In order to support a new BA, we just need to create the corresponding factory.
It should provide methods for creating a binding from the source code and parsing an input.

In order to use the library with a bunch of BAs, you need to define the corresponding nso<BAs...>, it would link everything toghether.